Remove #add_knight, add and change #make_tree
This commit removes #add_knight, as there is no need to actually move
the knight across the game board to find the shortest path.

Additionally, this commit changes #make_tree to #find_shortest_path, as
the method returns the knight that has the shortest path and location
that matches the dest after breadth-first search.