Refactor Knight to be Node-like, complete Board
This commit refactors Knight to be more node-like in that it tracks
children/parents, rather than tracking a history of locations due to
implementation issues in the last commit.

As a result, Knight also has a method #path that creates an array of the
moves from the root Knight to the current Knight.

Board changes as a result of Knight refactoring, and completes the
'#knight_moves' expectations, printing the path taken and the number of
moves required.